requirements.txt
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
h11==0.13.0
idna==3.3
outcome==1.2.0
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
schedule==1.1.0
selenium==4.3.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==4.3.0
urllib3==1.26.10
wsproto==1.1.0
